auto flattening and data processing broken out
